LayoutTests/fast/text-autosizing/ios/idempotentmode/idempotent-autosizing-skip-expected.txt

PASS result is >= 13
PASS result is >= 13
PASS result is 12
PASS result is 12
PASS result is 12
PASS result is 12
PASS result is >= result2
PASS result is >= 13
PASS successfullyParsed is true

TEST COMPLETE
Test
Test
Test
Test
Test
TestTestTestTest
Test
